[BE] refactor#179: 로그아웃 예외 정상 진행 처리
:tada: :tada: :tada:
이현수 <znight1020@naver.com>님이 Closes #179을 완료했습니다. 

커밋 목록
* refactor: 로그아웃 시 재발급 토큰이 유효하지 않다면 정상 진행

* refactor: 매니저 로그인 nickname 계정 생성

승인자
Approved-by: 송재원 <fosong98@gmail.com>

+++ module-api/src/main/java/com/e205/auth/jwt/handler/JwtRequestProcessHandler.java
@@ -47,13 +47,10 @@ public class JwtRequestProcessHandler {
   public void logout(HttpServletRequest request) {
     String headerValue = request.getHeader(REFRESH_TOKEN_HEADER_NAME);
     String refreshToken = extractRefreshToken(headerValue);
-
-    validateRefreshToken(refreshToken);
-
     Integer memberId = jwtProvider.getMemberId(refreshToken);
-    validateRefreshTokenInRedis(memberId);
-
-    jwtRepository.deleteRefreshTokenByMemberId(memberId);
+    if (jwtRepository.existByMemberId(memberId)) {
+      jwtRepository.deleteRefreshTokenByMemberId(memberId);
+    }
   }
 
   private String extractRefreshToken(String headerValue) {
